+++ main/transfer.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"kd.explorer/config"
-	"kd.explorer/service/kd"
+	"kd.explorer/service"
 	"flag"
 	"time"
 	"fmt"
@@ -13,9 +13,9 @@ func main() {
 	var t string
 	flag.StringVar(&config.CurUser, "u", "zhoushan_5781", "current user")
 	flag.StringVar(&t, "t", "", "sleep time")
-	flag.Float64Var(&config.SecKillFee, "fee", kd.SecKillMaxFEE, "")
-	flag.Float64Var(&config.SecKillRate, "rate", kd.SecKillMinRATE, "")
-	flag.IntVar(&config.SecKillRestDay, "rest", kd.SecKillMaxRestDAY, "")
+	flag.Float64Var(&config.SecKillFee, "fee", service.SecKillMaxFEE, "")
+	flag.Float64Var(&config.SecKillRate, "rate", service.SecKillMinRATE, "")
+	flag.IntVar(&config.SecKillRestDay, "rest", service.SecKillMaxRestDAY, "")
 	flag.Parse()
 
 	st := 1
@@ -25,7 +25,7 @@ func main() {
 
 	for {
 		// run analyse
-		kd.RunTA()
+		service.RunTA()
 
 		time.Sleep(time.Duration(st) * time.Second)
 		fmt.Println(fmt.Sprintf("sleep %d second", st))
