+++ service/rule.go
@@ -1,4 +1,4 @@
-package kd
+package service
 
 import (
 	"reflect"
